Slope in regions is now cuted out from slope map from the hole data set and not for each regions. Should be faster
